feat: extract RUM/web-traffic/CVE routes to routes/rum.py
