+++ templates/cve.html
@@ -127,7 +127,7 @@
           </div>
           <div class="d-flex gap-1" style="flex: 0 0 auto">
             <button type="submit" class="btn btn-primary btn-sm"><i class="bi bi-search"></i></button>
-            <a href="{{ url_for('view_enrichment_cve') }}" class="btn btn-outline-secondary btn-sm"><i class="bi bi-x"></i></a>
+            <a href="{{ url_for('rum.view_enrichment_cve') }}" class="btn btn-outline-secondary btn-sm"><i class="bi bi-x"></i></a>
           </div>
         </form>
 {% endcall %}
@@ -230,7 +230,7 @@
   <i class="bi bi-shield-check" style="font-size:2rem;"></i>
   <p class="mt-2">
     {% if severity_filter or ecosystem_filter or package_filter %}
-    No findings match the current filters. <a href="{{ url_for('view_enrichment_cve') }}">Clear filters</a>
+    No findings match the current filters. <a href="{{ url_for('rum.view_enrichment_cve') }}">Clear filters</a>
     {% else %}
     No CVE findings stored yet.
     {% if cve_last_scan %}Last scan: <span class="sobs-tz-ts" data-utc-ts="{{ cve_last_scan }}">{{ cve_last_scan[:19] }}</span>.{% endif %}
@@ -245,8 +245,8 @@
 
 {% block scripts %}
 <script>
-var librariesApiUrl = {{ url_for('api_enrichment_libraries') | tojson }};
-var repoHealthApiUrl = {{ url_for('api_enrichment_github_repo_health') | tojson }};
+var librariesApiUrl = {{ url_for('rum.api_enrichment_libraries') | tojson }};
+var repoHealthApiUrl = {{ url_for('rum.api_enrichment_github_repo_health') | tojson }};
 var cveScanUrl = {{ url_for('api_cve_scan') | tojson }};
 var cveDispositionBaseUrl = {{ url_for('api_cve_set_disposition', osv_id='__OSV_ID__') | tojson }};
