+++ templates/cve_help.html
@@ -17,7 +17,7 @@
             <p class="text-secondary mb-0">Library vulnerability detection, OSV.dev enrichment, and manual scan triggers.</p>
         </div>
         <a class="btn btn-outline-secondary btn-sm"
-           href="{{ url_for('view_enrichment_cve') }}"
+           href="{{ url_for('rum.view_enrichment_cve') }}"
            title="Back to CVE Findings">
             <i class="bi bi-arrow-left me-1"></i><span class="help-btn-label">Back to CVE Findings</span>
         </a>
