+++ templates/rum_help.html
@@ -17,7 +17,7 @@
             <p class="text-secondary mb-0">Real User Monitoring session replay, performance data, and browser event browsing.</p>
         </div>
         <a class="btn btn-outline-secondary btn-sm"
-           href="{{ url_for('view_rum') }}"
+           href="{{ url_for('rum.view_rum') }}"
            title="Back to RUM">
             <i class="bi bi-arrow-left me-1"></i><span class="help-btn-label">Back to RUM</span>
         </a>
