+++ templates/settings.html
@@ -235,7 +235,7 @@
                         daily CVE scanning (OSV.dev, Apache 2.0, free).
                         Library versions are auto-detected from OTEL <code>telemetry.sdk.*</code> attributes.
                         Findings are stored in <code>sobs_cve_findings</code> and shown on the
-                        <a href="{{ url_for('view_web_traffic') }}">Web Traffic</a> page.
+                        <a href="{{ url_for('rum.view_web_traffic') }}">Web Traffic</a> page.
                     </p>
                     <div class="d-flex align-items-center justify-content-end mt-2">
                         <a href="{{ url_for('settings.view_enrichment_settings') }}"
